requirements.txt
numpy==2.1.3
pdf2image==1.17.0
pillow==11.0.0
